# Conflicts:
#	Players.txt
#	out/production/Color_Switch/Login_Menu.class
#	src/Login_Menu.java
#	src/Player_Menu.java

+++ src/Login_Menu.java
@@ -5,6 +5,7 @@ import javafx.fxml.Initializable;
 import javafx.geometry.Point3D;
 import javafx.scene.Group;
 import javafx.scene.control.Button;
+import javafx.scene.media.AudioClip;
 import javafx.scene.transform.Rotate;
 
 import java.io.*;
@@ -12,7 +13,7 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.ResourceBundle;
 
-public class Login_Menu implements Initializable {
+public class Login_Menu implements Initializable,Serializable {
 
     @FXML
     private Group obs1;
@@ -29,7 +30,6 @@ public class Login_Menu implements Initializable {
     @FXML
     private Group t2;
 
-    static ArrayList<Player> Plist=new ArrayList<>();
     static ObjectInputStream in;
     static ObjectOutputStream out;
 
@@ -92,37 +92,41 @@ public class Login_Menu implements Initializable {
     }
 
 
-    public void exit(ActionEvent e){System.exit(0);}
+    public void exit(ActionEvent e){
+        URL path = getClass().getResource("/soundeffects/button.wav");
+        AudioClip ac = new AudioClip(path.toString());
+        ac.play();
+        // for delay in threading
+        for(int i=0;i<1000000000;i++){;}
+        System.exit(0);
+    }
 
     public void Exiting_Player(ActionEvent e) throws IOException, ClassNotFoundException {
-
-
-        ArrayList<Player> temp =new ArrayList<>();
-        Player p;
-
-        in=new ObjectInputStream(new FileInputStream("Players.txt"));
-        try {
-            while (true) {
-                p = (Player) in.readObject();
-                temp.add(p);
-                System.out.println(p.getName());
-            }
-        }
-        catch(Exception e1)
-        {}
-
-
+        URL path = getClass().getResource("/soundeffects/button.wav");
+        AudioClip ac = new AudioClip(path.toString());
+        ac.play();
+        Frame.navigation.load("Existing_Player_Details.fxml");
     }
 
     public static void addPlayer(Player p) throws IOException {
-        out=new MyObjectOutputStream(new FileOutputStream("Players.txt",true));
+
+        System.out.println("game is saving" + p.getName());
+        File f = new File("Players.txt");
+        if(f.isFile()){
+            out=new MyObjectOutputStream(new FileOutputStream("Players.txt",true));
+        }
+        else {
+            out = new ObjectOutputStream(new FileOutputStream("Players.txt", true));
+        }
         out.writeObject(p);
-        Plist.add(p);
+
     }
 
 
     public void New_Player(ActionEvent e){
-
+        URL path = getClass().getResource("/soundeffects/button.wav");
+        AudioClip ac = new AudioClip(path.toString());
+        ac.play();
         Frame.navigation.load("New_Player_Details.fxml");
     }
